Moved libsla unit tests
