+++ crates/libsla/src/tests.rs
@@ -1,0 +1,2 @@
+mod opcodes;
+mod sleigh;
